chore: moved assets to app folder
